+++ src/main/java/com/yanny/ytech/registration/YTechItems.java
@@ -215,19 +215,23 @@ public class YTechItems {
     }
 
     private static Item helmetItem(MaterialType material) {
-        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.HELMET, new Item.Properties());
+        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.HELMET,
+                new Item.Properties().durability(ArmorItem.Type.HELMET.getDurability(YTechArmorMaterials.ARMORS.getDurability(material))));
     }
 
     private static Item chestplateItem(MaterialType material) {
-        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.CHESTPLATE, new Item.Properties());
+        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.CHESTPLATE,
+                new Item.Properties().durability(ArmorItem.Type.CHESTPLATE.getDurability(YTechArmorMaterials.ARMORS.getDurability(material))));
     }
 
     private static Item leggingsItem(MaterialType material) {
-        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.LEGGINGS, new Item.Properties());
+        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.LEGGINGS,
+                new Item.Properties().durability(ArmorItem.Type.LEGGINGS.getDurability(YTechArmorMaterials.ARMORS.getDurability(material))));
     }
 
     private static Item bootsItem(MaterialType material) {
-        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.BOOTS, new Item.Properties());
+        return new ArmorItem(YTechArmorMaterials.ARMORS.get(material), ArmorItem.Type.BOOTS,
+                new Item.Properties().durability(ArmorItem.Type.BOOTS.getDurability(YTechArmorMaterials.ARMORS.getDurability(material))));
     }
 
     private static Item hoeItem(MaterialType material) {